third_party/blink/web_tests/platform/generic/http/tests/fullscreen/fullscreenelement-different-origin-expected.txt

Test iframe from different origin can not access webkitFullscreenElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.webkitFullscreenElement is targetDiv1
PASS document.webkitCurrentFullScreenElement is targetDiv1
PASS message is "inner-iframe.html document.webkitFullscreenElement = null"
PASS message is "inner-iframe.html document.webkitCurrentFullScreenElement = null"
PASS successfullyParsed is true

TEST COMPLETE
